Revert "Move jit/type and utils/functional to ATen/core (#12466)"
This reverts commit f989d4b18e342121845ac8f370368a7d92f7611b.
