+++ torch/csrc/jit/type.cpp
@@ -1,4 +1,6 @@
-#include <ATen/core/jit_type.h>
+#include "torch/csrc/jit/type.h"
+
+#include "torch/csrc/jit/assertions.h"
 
 #include <iostream>
 
@@ -9,7 +11,7 @@ std::ostream& operator<<(std::ostream & out, const Type & t) {
     out << at::toString(value->scalarType()) << "(";
     auto& sizes = value->sizes();
     auto& strides = value->strides();
-    AT_ASSERT(sizes.size() == strides.size());
+    JIT_ASSERT(sizes.size() == strides.size());
     for (size_t i = 0; i < sizes.size(); i++) {
       if (i > 0) {
         out << ", ";
@@ -239,7 +241,7 @@ TypePtr matchTypeVariables(TypePtr formal, TypePtr actual, TypeEnv& type_env) {
 }
 
 // change return types like List[List[t]] into List[List[int]]
-CAFFE2_API TypePtr evalTypeVariables(TypePtr type, std::unordered_map<std::string, TypePtr>& type_env) {
+TORCH_API TypePtr evalTypeVariables(TypePtr type, std::unordered_map<std::string, TypePtr>& type_env) {
   if(!type->hasFreeVariables())
     return type;
